partially working codegen
