+++ assembler/asm.h
@@ -3,17 +3,24 @@
 
 #include <assert.h>
 #include <ctype.h>
-#include <stdio.h>
+#include <cstdio>
 #include <stdlib.h>
 
-#include "../commands.h"
+#include "../enums.h"
 
 const size_t       RUNS_CNT       = 2;
 const size_t       RUN_LBL_UPD    = 1;
 
 const size_t       MAX_LINES      = 100;
 const size_t       CMDS_PER_LINE  = 2;
+
 const char * const DFLT_CMDS_FILE = "user_commands.txt";
+
+struct Label
+{
+    int    cmd_ptr;
+    char * name;
+};
 
 typedef enum {
     ASM_OUT_NO_ERR = 0,
