+++ processor/spu.cpp
@@ -5,12 +5,12 @@
 #include <unistd.h>
 #include <math.h>
 
-#include "../commands.h"
+#include "../enums.h"
 #include "spu.h"
 #include "../stacklib/stack.h"
 #include "../text_processing_lib/text_buf.h"
 
-const int SHOW_INTERMED_INFO = 0;
+#define SHOW_INTERMED_INFO 0
 #define PRINTF_INTERMED_INFO(format, ...)     \
     if (SHOW_INTERMED_INFO)                   \
         fprintf(stderr, format, __VA_ARGS__); \
@@ -108,369 +108,20 @@ RunBinRes RunBin (const cmd_code_t * prog_code, size_t n_bytes, SPU * spu)
 
         switch (cmd & OPCODE_MSK)
         {
-            case CMD_HLT:
-            {
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Hlt encountered, goodbye!\n", "proc", ip);
-
-                return REACH_HLT;
-            }
-
-            case CMD_FRAME:
-            {
-                for (int i = 0; i < SPU_GRAM_HIGHT; i++, printf("\n"))
-                {
-                    for (int j = 0; j < SPU_GRAM_WIDTH; j++)
-                    {
-                        printf("%c%c", spu->RAM[GRAM_MAPPING + i * SPU_GRAM_HIGHT + j] / STK_PRECISION,
-                                       spu->RAM[GRAM_MAPPING + i * SPU_GRAM_HIGHT + j] / STK_PRECISION);
-                    }
-                }
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_PUSH:
-            {
-                Elem_t arg = GetPushArg(prog_code, ip, spu->gp_regs, spu->RAM);
-
-                PushStack(&spu->stk, arg);
-
-                ip += CalcIpOffset(cmd);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Push GetArg -> %d\n", "proc", ip_init, arg);
-
-                break;
-            }
-
-            case CMD_POP:
-            {
-                if (cmd & ARG_TYPE_MSK)
-                {
-                    Elem_t * arg_ptr = GetPopArgAddr(prog_code, ip, spu->gp_regs, spu->RAM);
-
-                    if (arg_ptr == NULL)
-                    {
-                        fprintf(stderr, "Processor Error! SetArg couldnt return stuff\n");
-                        abort();
-                    }
-
-                    pop_err = POP_NO_ERR;
-
-                    *arg_ptr = PopStack(&spu->stk, &pop_err);
-
-                    ip += CalcIpOffset(cmd);
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Pop number to %p\n", "proc", ip_init, arg_ptr);
-                }
-                else
-                {
-                    pop_err = POP_NO_ERR;
-                    PopStack(&spu->stk, &pop_err);
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Pop number\n", "proc", ip_init);
-
-                    ip += CalcIpOffset(cmd);
-                }
-
-                break;
-            }
-
-            case CMD_IN:
-            {
-                fprintf(stdout, "\n>> ");
-
-                fscanf(stdin, "%d", &val);
-
-                val *= STK_PRECISION;
-
-                PushStack(&spu->stk, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_OUT:
-            {
-                pop_err = POP_NO_ERR;
-
-                val = PopStack(&spu->stk, &pop_err);
-
-                fprintf(stdout, "\n<< %.2f\n", (float) val / STK_PRECISION);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_ADD:
-            {
-                // PUSH(POP() + POP())
-                pop_err = POP_NO_ERR;
-                val = PopStack(&spu->stk, &pop_err) + PopStack(&spu->stk, &pop_err);
-                PushStack(&spu->stk, val);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Add: %d\n", "proc", ip_init, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_SUB:
-            {
-                pop_err = POP_NO_ERR;
-                val -= PopStack(&spu->stk, &pop_err);
-                val += PopStack(&spu->stk, &pop_err);
-                PushStack(&spu->stk, val);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Sub: %d\n", "proc", ip_init, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_MUL:
-            {
-                pop_err = POP_NO_ERR;
-                val = MultInts(PopStack(&spu->stk, &pop_err), PopStack(&spu->stk, &pop_err));
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Mul: %d\n", "proc", ip_init, val);
-
-                PushStack(&spu->stk, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_DIV:
-            {
-                pop_err = POP_NO_ERR;
-
-                Elem_t denominator = PopStack(&spu->stk, &pop_err);
-                Elem_t numerator   = PopStack(&spu->stk, &pop_err);
-
-                val = 0;
-                val = DivideInts(numerator, denominator);
-
-                PushStack(&spu->stk, val);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Div: %d\n", "proc", ip_init, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_SQRT:
-            {
-                pop_err = POP_NO_ERR;
-                val = PopStack(&spu->stk, &pop_err);
-
-                val = (int) sqrt(val * STK_PRECISION);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Sqrt: %d\n", "proc", ip_init, val);
-
-                PushStack(&spu->stk, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_SQR:
-            {
-                pop_err = POP_NO_ERR;
-                val = PopStack(&spu->stk, &pop_err);
-
-                val = val * val / STK_PRECISION;
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Sqr: %d\n", "proc", ip_init, val);
-
-                PushStack(&spu->stk, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_MOD:
-            {
-                pop_err = POP_NO_ERR;
-                Elem_t denominator = PopStack(&spu->stk, &pop_err);
-                Elem_t numerator   = PopStack(&spu->stk, &pop_err);
-
-                val = CalcMod(numerator, denominator);
-
-                PushStack(&spu->stk, val);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Mod: %d\n", "proc", ip_init, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_IDIV:
-            {
-                pop_err = POP_NO_ERR;
-                Elem_t denominator = PopStack(&spu->stk, &pop_err);
-                Elem_t numerator   = PopStack(&spu->stk, &pop_err);
-
-                val = CalcIdiv(numerator, denominator);
-
-                PushStack(&spu->stk, val);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Idiv: %d\n", "proc", ip_init, val);
-
-                ip += CalcIpOffset(cmd);
-
-                break;
-            }
-
-            case CMD_CALL:
-            {
-                PushStack(&spu->call_stk, (Elem_t)(ip + sizeof(cmd_code_t) + sizeof(Elem_t)));
-
-                ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Call to %lu\n", "proc", ip_init, ip);
-
-                break;
-            }
-
-            case CMD_RET:
-            {
-                pop_err = POP_NO_ERR;
-                ip = PopStack(&spu->call_stk, &pop_err);
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Ret to %lu\n", "proc", ip_init, ip);
-
-                break;
-            }
-
-            case CMD_JMP:
-            {
-                ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-
-                break;
-            }
-            case CMD_JA:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res > 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
-            case CMD_JAE:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res >= 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
-
-            case CMD_JB:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res < 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
-
-            case CMD_JBE:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res <= 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
-
-            case CMD_JE:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res == 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
-
-            case CMD_JNE:
-            {
-                int cmp_res = PopCmpTopStack(&spu->stk);
-
-                if (cmp_res != 0)
-                {
-                    ip = *(Elem_t *)(prog_code + ip + sizeof(cmd_code_t));
-
-                    PRINTF_INTERMED_INFO("# (%s - %3ld) Jmp to %lu\n", "proc", ip_init, ip);
-                }
-                else
-                {
-                    ip += CalcIpOffset(cmd); // skip integer pointer to a position in code
-                }
-
-                break;
-            }
+            #define DEF_CMD(name, num, text, spu_code, ...)     \
+                case CMD_##name:                                \
+                    {                                           \
+                        spu_code;                               \
+                        break;                                  \
+                    }                                           \
+
+            #include "../commands.h"
+
+            #undef DEF_CMD
 
             default:
             {
-                fprintf(stderr, "SIGILL! Illegal instruction \"%d\" (%lu)\n", cmd, ip);
+                fprintf(stderr, "SIGILL! Illegal instruction \"%d\" (%lu)\n", cmd, ip_init);
 
                 return ILL_CDMCODE;
             }
